Marvel Collection App
This application was written to mimic the android Neko application, In which a cat would be added to the SQL lite database,
for every notification sent to the device. The notification listener was used, along with data stored within SQL lite using 
cupboard to parse the data. In this particular application, a marvel character is to be unlocked, every time
a notification update is opened. The Marvel character would display unique history and collected within the sql lite database
as a collection application. An added feature, I wanted to implement a unique activity where each marvel character will have a 
unique workout plan to keep the character unlocked. I wanted to create a useful application where The application is a combination
of a daily activity routine along with a collection database of marvel characters.. which will be unlocked over a course of 30 days
with a 30 day exercise plan.